requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-MySQLdb==2.0.0
gunicorn==23.0.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.7
packaging==25.0
python-dotenv==1.1.0
Werkzeug==3.1.3
zipp==3.21.0